Automation-Android/MXPlayerUITest/requirements.txt
aiofiles==0.7.0
Appium-Python-Client==1.0.2
atomicwrites==1.4.0
attrdict==2.0.1
attrs==20.3.0
backports.entry-points-selectable==1.1.0
boto==2.49.0
click==8.0.1
colorama==0.4.4
distlib==0.3.3
execnet==1.9.0
fb-idb==1.1.5
filelock==3.0.12
Flask==2.0.1
future==0.18.2
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.1
importlib-metadata==2.0.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
logit==0.7.4
MarkupSafe==2.0.1
packaging==20.4
pbr==5.6.0
platformdirs==2.3.0
pluggy==0.13.1
protobuf==3.18.0
py==1.9.0
pycrypto==2.6.1
pyparsing==2.4.7
pytest==6.1.2
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==10.2
pytest-testconfig==0.2.0
pytest-xdist==2.4.0
PyYAML==5.4.1
selenium==3.141.0
six==1.15.0
stevedore==3.4.0
toml==0.10.2
treelib==1.6.1
urllib3==1.25.11
virtualenv==20.8.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
Werkzeug==2.0.1
zipp==3.4.0
